fix required size returned by WideCharToMultiByte on linux
WideCharToMultiByte returned the size without terminating zero IF that size is
exactly equal to the number of wide characters including terminating zero.

This is the case if the source contains exactly one wide char that is converted into a
two-byte utf-8 sequence.

This then caused an assertion in DxcGetBlobAsUtf8:
"otherwise, CodePageBufferToUtf8 failed to null-terminate buffer."